+++ lib/core/utils/index.js
@@ -53,7 +53,7 @@ export {
   isLabelledShadowDomSelector,
   isLabelledFramesSelector
 } from './is-context';
-export { default as couldBeCustomElement } from './could-be-custom-element';
+export { default as isValidCustomElementName } from './is-valid-custom-element-name';
 export { default as isHidden } from './is-hidden';
 export { default as isHtmlElement } from './is-html-element';
 export { default as isNodeInContext } from './is-node-in-context';
