+++ lib/core/utils/is-valid-custom-element-name.js
@@ -1,5 +1,3 @@
-import isXHTML from './is-xhtml';
-
 // reserved names that match the custom element regex but are not valid custom elements
 // @see https://html.spec.whatwg.org/multipage/custom-elements.html#valid-custom-element-name
 const reservedNames = [
@@ -20,15 +18,13 @@ const validLocalNameRegex =
 
 /**
  * Verifies that a given node could be a custom-element (without testing the CustomElement registry there's no way to know for sure)
- * @method couldBeCustomElement
+ * @method isValidCustomElementName
  * @memberof axe.utils
- * @param {HTMLElement} node node to check
+ * @param {string} nodeName nodeName to validate
  * @return {Boolean} true/ false
  */
-export default function couldBeCustomElement(node) {
-  const nodeName = isXHTML(node.ownerDocument)
-    ? node.nodeName
-    : node.nodeName.toLowerCase();
+export default function isValidCustomElementName(nodeName) {
+  nodeName = nodeName.toLowerCase();
 
   // @see https://html.spec.whatwg.org/#valid-custom-element-name
   return (
